Move audit logger config
