+++ core/config/app_config.go
@@ -6,8 +6,6 @@ import (
 	"github.com/google/uuid"
 	"github.com/pkg/errors"
 	"go.uber.org/zap/zapcore"
-
-	"github.com/smartcontractkit/chainlink/v2/core/logger/audit"
 )
 
 // nolint
@@ -50,7 +48,7 @@ type AppConfig interface {
 	Web
 
 	Database() Database
-	AuditLogger() audit.Config
+	AuditLogger() AuditLogger
 	Keeper() Keeper
 	TelemetryIngress() TelemetryIngress
 	Sentry() Sentry
